Add files via upload
In the second update, more advanced implementation has been done such as reversing from the central node, loop detection and so on.
The class has now been changed to the singleton class, the node class. You need to make a new LinkList object for every such usage and it seems appropriate to do so.